py/obj: Reuse mp_obj_get_int_maybe() in mp_obj_get_int().
The maybe() variant was added in 8270e3853dc167d, and get_int() was already there.
I didn't see any mention as to why it was cloned, so I guess it went unnoticed, or
that I missed something.

mp_obj_get_int() is used extensively so it's debatable whether this change, which
incurs a tiny code decrease, is worth the slight performance hit.